% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualComposite.R
\name{annualComposite}
\alias{annualComposite}
\title{Annual composite from a time series RasterBrick}
\usage{
annualComposite(x, fun, period = NULL, dates = NULL, years = NULL,
  sensor = NULL, na.rm = NULL, ...)
}
\arguments{
\item{x}{RasterBrick or RasterStack}

\item{fun}{Function to apply over each pixel for each period}

\item{period}{Numeric. Optional: vector of julian days to limit the calculation. Can be used to restrict compositing to a specific season, for example.}

\item{dates}{Date. Optional: vector of dates exactly corresponding to the layers of x. If not included, dates must be included in the z dimension of x (see \code{\link{getZ}}) or in \code{names(x)}.}

\item{years}{Numeric. Optional: Vector of years to which to limit the composite.}

\item{sensor}{Character. Optional: limit calculation to images from a particular sensor. Defaults to "all", but can take any of "TM", "ETM+", "ETM+ SLC-off" or "ETM+ SLC-on". Will be ignored with a warning if \code{names(x)} do not correspond to Landsat scene ID's.}

\item{...}{Arguments to be passed to \code{\link{mc.calc}}}
}
\value{
RasterBrick with results of \code{fun} for each year represented in the input time series RasterBrick.
}
\description{
Generates annual composites based on a time series RasterBrick, with the option of limiting to a specific period (e.g. season) in the year.
}
\details{
If \code{fun} takes a \code{na.rm} argument and none is supplied, the default value for \code{na.rm} for that function will be used.
}
\examples{
# load tura RasterBrick
data(tura)

# calculate mean and standard deviation values per year for ETM+ data only
annualMean <- annualSummary(tura, fun=mean, na.rm=TRUE, sensor="ETM+")
plot(annualMean)
annualSD <- annualSummary(tura, fun=sd, na.rm=TRUE, sensor="ETM+")
plot(annualSD)

# custom function to calculate # of non-NA values per pixel per year (similar to countObs())
ff <- function(x)
 length(x[!is.na(x)])
annualObs <- annualSummary(tura, fun=ff, sensor="ETM+")
}
\author{
Ben DeVries
}
\seealso{
\code{\link{summaryBrick}}
}

