\name{bfastSpatial-package}
\alias{bfastSpatial-package}
\alias{bfastSpatial}
\docType{package}
\title{
Utilities for change monitoring on spatio-temporal data
}
\description{
Pre-process gridded time-series data in order for them to be analyzed with change detection algorithms such as bfast. Uses classes from the raster package and includes utilities to run the algorithms and post-process the results.
}
\details{
\tabular{ll}{
Package: \tab bfastSpatial\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-06\cr
License: \tab tbd\cr
}

}
\author{
Loic Dutrieux, Ben DeVries

Maintainer: <loic.dutrieux@gmail.com>
}
\references{
Add publications here
}

\keyword{ Landsat }
\seealso{
bfast
}
\examples{
Add exemples here
}
