% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitToInt.R
\name{bitToInt}
\alias{bitToInt}
\title{Bit index to numeric}
\usage{
bitToInt(x)
}
\arguments{
\item{x}{Vector of numeric where each value correspond to a bit position}
}
\value{
Numeric
}
\description{
Converts a vector of bit positions to the correspondng integer. This aim of this function is to help users dealing with bit coded information, such as Pixel Quality information provided with Landsat and MODIS data.
}
\examples{
# We want to know the base 10 numeric value of a binary word that has bit indices 0, 4 and 5 activated (0011 0001)
bitToInt(c(0, 4, 5))
}
\author{
Loic Dutrieux
}
