% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bpPheno.R
\name{bpPheno}
\alias{bpPheno}
\title{Detect breakpoints in irregularly spaced time-series containing seasonal and trend components}
\usage{
bpPheno(x, order = 1, formula = response ~ trend + harmon, breaks = NULL,
  h = 0.15, nbreaks = NULL, mc.cores = 1)
}
\arguments{
\item{x}{time-series object of class zoo}

\item{order}{Numeric Harmonic order of the regression model}

\item{formula}{See \link{breakpoints}}

\item{breaks}{See \link{breakpoints}}

\item{h}{See \link{breakpoints}}

\item{nbreaks}{Numeric, maximum theoretical number of breaks (for spatial implementation)}
}
\description{
Detect breakpoints in irregularly spaced time-series containing seasonal and trend components
}

