% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bpPhenoSp}
\alias{bpPhenoSp}
\title{Runs breakpoints for a spatial subset with aggregation}
\usage{
bpPhenoSp(x, y, formula = response ~ trend + harmon, order = 1, h = 0.1,
  breaks = NULL, mc.cores = 1, ...)
}
\arguments{
\item{x}{A rasterBrick or rasterStack, ideally with time written to the z dimension. In case time is not written to the z dimension, the \code{dates=} argument has to be supplied (see \link{zooExtract})}

\item{y}{A SpatialPoints, SpatialPointsDataFrame, SpatialPolygons, SpatialPolygonsDataFrame, SpatialLines, SpatialLinesDataFrame, or extent. \link{breakpoints} will be ran at these locations. In case each feature of the object covers several pixels (typically SpatialPolygons(DataFrames), SpatialLines(DataFrames) and extent), an aggregation function (\code{fun=}) has to be supplied (see \link{extract}).}

\item{formula}{See \code{\link{breakpoints}}}

\item{order}{See \code{\link{bfastpp}}}

\item{h}{See \code{\link{breakpoints}}}

\item{breaks}{See \code{\link{breakpoints}}}

\item{mc.cores}{Numeric NUmber of cores to use (for parallel processing)}

\item{...}{Arguments to be passed to \link{zooExtract}}
}
\description{
Runs \link{breakpoints} on a rasterBrick object for a set of locations, determined by an object of class \link{Spatial-class}.
}
\examples{
# Load data
data(tura)

# 1- SpatialPoints case
# Generate SpatialPoints
sp <- sampleRegular(x = tura, size = 20, sp=TRUE)

# Run for each pixel overlapped by a point
out <- bpPhenoSp(tura, sp, formula = response ~ trend, h = 0.1)

# Visualize the results
plot(tura, 166)
plot(out, pch=16, cex = out$nbreaks, add = TRUE)

# 2 - SpatialPolygons case
data(turaSp)

out2 <- bpPhenoSp(tura, turaSp, fun = mean, formula = response ~ trend, h = 0.1)
# Visualize
plot(tura, 166)
# Build color palette
colfunc <- colorRampPalette(c("yellow", "red"))
colList <- colfunc(2008 - 1987)
plot(out2, col = colList[floor(out2$break1) - 1986], add = TRUE)
# Interpretation: The redder the latter the first break was detected. If transparent, no break detected in spatially aggregated polygon time-series.
}
\author{
Loic Dutrieux
}

