% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isLandsatSceneID.R
\name{.isLandsatSceneID}
\alias{.isLandsatSceneID}
\title{Check if a String is a Landsat Scene ID}
\usage{
.isLandsatSceneID(x)
}
\arguments{
\item{x}{Character or object of class RasterBrick or RasterStack}
}
\value{
\code{TRUE} if all elements of \code{x} match Landsat scene ID criteria, or \code{FALSE} otherwise.
}
\description{
Parses a vector of characters to check if they conform to Landsat naming convention
}
\author{
Ben DeVries, Loic Dutrieux
}
