% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitHarmonTrend.R
\name{fitHarmonTrend}
\alias{fitHarmonTrend}
\title{Harmonic-trend fit}
\usage{
fitHarmonTrend(x, dates = NULL, formula = response ~ harmon + trend,
  order = 1, pptype = "irregular", start = NULL, end = NULL,
  mc.cores = 1, sensor = NULL, ...)
}
\arguments{
\item{x}{raster times series}

\item{formula}{formula to fit to time series}

\item{order}{order of the harmonic trend (leave as 1 if there is no harmonic component)}

\item{start}{The start of the subset to be fit. See 'details' for special cases}

\item{end}{The end of the subset to be fit. See 'details' for special cases.}
}
\description{
Fit harmonic and/or trend model to a raster time series, or temporal subset thereof.
}
\details{
If \code{start} and \code{end} are left as \code{NULL}, the model will be fit to the entire time series. A \code{numeric} can be supplied to either or both arguments to subset the time series as in \code{\link{window}}.
}
