% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getMODISinfo.R
\name{getMODISinfo}
\alias{getMODISinfo}
\title{Get MODIS infos}
\usage{
getMODISinfo(x)
}
\arguments{
\item{x}{Character. MODIS file name}
}
\value{
A dataframe with one column ($date) of class \code{Date}
}
\description{
Parses through MODIS file names and retrieve information on acquisition date. Future versions will include retrieval of sensor information, tile ID, etc... The function is vectorized.
}
\examples{
# Simple case
fileName <- '/path/to/file/MOD17A2.A2008097.h11v09.005.2008142064019.hdf'
getMODISinfo(fileName)

# Multile files
fileNames <- c('/path/to/file/MOD17A2.A2008097.h11v09.005.2008142064019.hdf', 'MOD17A2.A2008097.h11v09.005.2008164064019.tif')
getMODISinfo(fileNames)

# Get only a vector of dates
getMODISinfo(fileNames)$date
}
\author{
Loic Dutrieux
}

