\docType{data}
\name{LE72220622001176EDC01-LE72220622002147EDC00-LT52220622004289CUB01}
\alias{LE72220622001176EDC01-LE72220622002147EDC00-LT52220622004289CUB01}
\title{Landsat-like tarballs}
\description{
  subset of Landsat surface reflectance data for testing
  pre-processing steps.
}
\author{
  NASA/USGS
}
\keyword{data}
\keyword{datasets}

