% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr2vi.R
\name{sr2vi}
\alias{sr2vi}
\title{Calculate Vegetation Index from Landsat surface reflectance data file}
\usage{
sr2vi(x, vi = "ndvi", e = NULL, mask = NULL, keep = c(0), L = 0.5,
  ...)
}
\arguments{
\item{x}{Character. File name of the hdf file containing the bands, or list of filenames (geoTiffs).}

\item{vi}{Character. Vegetation index to be computed. See 'details' for a list of the supported indices. The function will first attempt to find if a pre-processed version of the index is present in the list supplied or in the .hdf file and will use that layer when available.}

\item{e}{Extent object or object that can be coerced as extent.}

\item{mask}{Character or NULL. The name of the cloud/Land mask to be applied to the output (e.g.: \code{mask='fmask'})}

\item{keep}{Numeric. Can take multiple values. Which values of the mask layer should be kept?}

\item{L}{Numeric. Soil-adjustment factor for SAVI (ignored if vi != 'savi'). L can take on values between 0 and 1, and a default of 0.5 is typically used.}

\item{...}{Arguments to be passed to \code{\link{writeRaster}}}
}
\value{
A rasterLayer object
}
\description{
Extract layers, apply mask (optional), crop (optional), calculate NDVI (possible other indices supported in the future) and write output to file (optional)
}
\details{
A number of indices can be passed to \code{vi} at the moment, includeing \code{ndvi}, \code{evi}, \code{savi}, \code{nbr}, \code{ndmi}, \code{ndwi} and \code{mndwi}. Three tasseled cap components - brightness, greenness and wetness - are also supported. For brightness, \code{tcb}, \code{brightness} and \code{tcbright} are all acceptable arguments for \code{vi}. The same holds true for the other 2 indices.

For tasseled cap components, note that at the moment only a single set of coefficients for surface reflectance are used (Crist, 1985). Be aware that this may cause issues if OLI data are being used with ETM+ and TM data due to radiometric differences between OLI and the other sensors.
}
\author{
Loic Dutrieux and Ben DeVries
}
\seealso{
\code{\link{processLandsat}} and \code{\link{processLandsatBatch}} for wrapper and wrapper/batcher functions
}

