% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetRasterTS.R
\name{subsetRasterTS}
\alias{subsetRasterTS}
\title{Trim raster time series}
\usage{
subsetRasterTS(x, sensor = NULL, minDate = NULL, maxDate = NULL, ...)
}
\arguments{
\item{x}{RasterBrick or RasterStack}

\item{sensor}{Character. Limit time series to specified sensor(s). Can take any combination of "ETM+ SLC-off", "ETM+ SLC-on", "ETM+", "TM", "OLI". Defaults to \code{NULL} (use all sensors).}

\item{minDate}{Numeric. Optional: minumum date (in format c(year, julian day)) before which all layers will be removed from the RasterBrickStack.}

\item{maxDate}{Numeric. Optional: maximum date (in format c(year, julian day)) after which all layers will be removed form the RasterBrickStack.}

\item{...}{Additional arguments to be passed to \code{\link{raster::subset}}}
}
\description{
Trim a time series RasterBrickStack based on a specified Landsat sensor or threshold date (year)
}
\details{
This function only supports Landsat data with associated scene ID's as \code{names(x)} at this time. Support for MODIS (and possibly other datasets) will follow in future versions.
}
\author{
Ben DeVries
}

