% Generated by roxygen2 (4.0.2): do not edit by hand
\name{timeStackMODIS}
\alias{timeStackMODIS}
\title{Creates a time stack of Landsat layers}
\usage{
timeStackMODIS(x, pattern = NULL, ...)
}
\arguments{
\item{x}{character. dir containing the files to be stacked or character list (the files). IN the former case, it is recommended to use the \code{pattern} argument}

\item{pattern}{See \link{list.files}}

\item{...}{Arguments to be passed to \link{writeRaster}. If specifying a filename, it is strongly recommended to also set a datatype.}
}
\description{
Creates a stack of MODIS layers, with time written to z dimension. Works in anolog way to \link{\code{timeStack}}
}
\author{
Loic Dutrieux
}

