% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zooExtract.R
\name{zooExtract}
\alias{zooExtract}
\title{Utility to extract a zoo object from a raster time-series}
\usage{
zooExtract(x, sample = "click", dates = NULL, file = NULL, ...)
}
\arguments{
\item{x}{rasterBrick or rasterStack object, or file name to a multilayer raster object stored on disk.}

\item{sample}{The point(s) to extract. Default to 'click' where a point is interactively choosen by clicking. Or any of object accepted by \code{\link{extract}}. (points represented by a two-column matrix or data.frame, or \code{\link{SpatialPoints}*}; \code{\link{SpatialPolygons}*}; \code{\link[sp]{SpatialLines}}; \code{\link{Extent}}; or a numeric vector representing cell numbers). For 'click', a layer of \code{x} needs to be plotted first.}

\item{dates}{A date vector (optional, only if time is not yet contained in the z dimension of the raster object, or comprised in its layer names.)}

\item{file}{character rds filename where to write the output.}

\item{...}{Arguments to be passed to \link{extract}.}
}
\value{
A zoo object that may contain multiple time-series
}
\description{
This function extracts a zoo object from a multilayer raster object with time written to the z dimension or a date vector supplied externally. The aim of this utility is to facilitate the use of the bfmApp for bfastmonitor parameters investigation. The app allows the user to supply a zoo (time-series) object and visualize the effect of parameters change on the algorithm output. We outline in the vignette of the bfastSpatial package that this is an interesting step for parameter selection prior to more heavy processing when starting to work in a study area.
}
\examples{
# Single time-series selected interactively
data(tura)
plot(tura, 2)
zooExtract(x = tura, sample = 'click', file = file.path(tempdir(), 'zooClick.rds'))

# Then in order to explore the time series and the bfastmonitor parameters, run the following line
\dontrun{
runGitHub("bfmApp", "dutri001")
}
# And upload the file zooClick.rds
}
\author{
Loic Dutrieux
}
\seealso{
\link{sr2vi}
}

