% Generated by roxygen2 (4.0.1): do not edit by hand
\name{annualSummary}
\alias{annualSummary}
\title{Annual summary of a time series RasterBrick}
\usage{
annualSummary(x, fun, dates = NULL, years = NULL, sensor = NULL,
  na.rm = NULL, ...)
}
\arguments{
\item{x}{RasterBrick or RasterStack}

\item{fun}{Function to apply over each pixel for each year}

\item{dates}{Date. Optional: vector of dates exactly corresponding to the layers of x. If not included, dates must be included in the z dimension of x (see \code{\link{getZ}}) or in \code{names(x)}}

\item{years}{Numeric. Optional: Vector of years to which to limit the summary.}

\item{sensor}{Character. Optional: limit calculation to images from a particular sensor. Defaults to "all", but can take any of "TM", "ETM+", "ETM+ SLC-off" or "ETM+ SLC-on". Will be ignored with a warning if \code{names(x)} do not correspond to Landsat scene ID's.}

\item{...}{Arguments to be passed to \code{\link{mc.calc}}}
}
\value{
RasterBrick with results of \code{fun} for each year represented in the input time series RasterBrick.
}
\description{
Calculates pixel-based statistics for every year represented by a time series RasterBrick
}
\details{
If \code{fun} takes a \code{na.rm} argument and none is supplied, the default value for \code{na.rm} for that function will be used.
}
\examples{
# load tura RasterBrick
data(tura)

# calculate mean and standard deviation values per year for ETM+ data only
annualMean <- annualSummary(tura, fun=mean, na.rm=TRUE, sensor="ETM+")
plot(annualMean)
annualSD <- annualSummary(tura, fun=sd, na.rm=TRUE, sensor="ETM+")
plot(annualSD)

# custom function to calculate # of non-NA values per pixel per year (similar to countObs())
ff <- function(x)
 length(x[!is.na(x)])
annualObs <- annualSummary(tura, fun=ff, sensor="ETM+")
}
\author{
Ben DeVries
}
\seealso{
\code{\link{summaryBrick}}
}

