% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSceneinfo.R
\name{getSceneinfo}
\alias{getSceneinfo}
\title{Retrieve Landsat info from filenames}
\usage{
getSceneinfo(sourcefile, ...)
}
\arguments{
\item{sourcefile}{Character. Filename of a landsat layer or dataset.}

\item{...}{Additional arguments to pass to \code{\link{write.csv}}.}
}
\value{
a \code{data.frame} with parsed scene information from Landsat scene names
}
\description{
Parses through typical Landsat filenames and retrieves information on sensor and acquisition date. Vectorized over \code{sourcefile}.
}
\examples{
getSceneinfo(c('ndvi.LC82300702014234.tar.gz', 'ndvi.LT52300702008234.tif', '/home/username/LC08_L1TP_019046_20170311_20170317_01_T1_sr_band4.tif'))

# Load tura
data(tura)
getSceneinfo(names(tura))

}
\author{
Ben DeVries and Loic Dutrieux
}
