% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{LE70230282011250EDC00_sub.zip-LE70230282015250EDC00_sub_fake.zip}
\alias{LE70230282011250EDC00_sub.zip-LE70230282015250EDC00_sub_fake.zip}
\title{Landsat-like zip archives}
\description{
subset of Landsat surface reflectance data for testing pre-processing steps.
}
\author{
NASA/USGS
}
\keyword{data}

