% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.calc.R
\name{mc.calc}
\alias{mc.calc}
\title{Multicore implementation of the raster::\code{\link{calc}} function.}
\usage{
mc.calc(x, fun, mc.cores = 1, ...)
}
\arguments{
\item{x}{Raster* object}

\item{fun}{Function to be applied to the raster object.}

\item{...}{Arguments to be passed to \code{link{writeRaster}}.}
}
\value{
a Raster* object
}
\description{
Allows functions to be applied to raster objects, with multicore support.
}
\details{
For further help, see \code{\link{calc}}. Warnings of the parallel package (see \code{\link{mclapply}} for instance) apply to this function.
}
\seealso{
\code{\link{calc}}
}
\author{
Loic Dutrieux
}
