% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{tura}
\alias{tura}
\title{RasterBrick object}
\source{
http://earthexplorer.usgs.gov
}
\usage{
data(tura)
}
\description{
Time series raster brick for an area in Tura kebele, Kafa Zone, SW Ethiopia. Values are NDVI rescaled by a factor of 10000. Data originate from the Landsat 5 TM and Landsat 7 ETM+ sensors. Original Landsat scene names can be found by typing \code{names(tura)}. Dates are also contained in the z-dimension (\code{getZ(tura)}; see also \code{\link{timeStack}}).
}
\examples{
data(tura)

# scene names
names(tura)
s <- getSceneinfo(names(tura))
s$year <- as.numeric(format(s$date, format = "\%Y"))
hist(s$year, breaks = c(1980:2015))

# plot first 9 scenes
plot(tura, 1:9, main = getZ(tura))

# plot a time series of the 50th cell
plot(getZ(tura), tura[50], xlab="time", ylab="NDVI (x 10000)")

# make a new rasterBrick from only ETM+ scenes
x <- subsetRasterTS(tura, sensor = "ETM+")

# plot time series from the revised brick
plot(getZ(x), x[50], xlab="time", ylab="NDVI (x 10000)")
}
\references{
DeVries B., Verbesselt J., Kooistra L. and Herold M. (2015). Robust Monitoring of Small-Scale Forest Disturbances in an Afromontane Forest Using Landsat Time Series. Remote Sensing of Environment, 161:107-121.
}
\author{
Ben DeVries
}
\keyword{data}
