% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{turaSp}
\alias{turaSp}
\title{SpatialPolygons object}
\usage{
data(turaSp)
}
\description{
Three polygons located in the tura area (ethiopia)
}
\examples{
data(turaSp)
data(tura)
plot(tura, 2)
plot(turaSp, add = TRUE)
}
\author{
Ben DeVries
}
\keyword{data}

